Updating my biography
To create my ReadMe biography, I cloned the forked repo to my desktop using GitKraken. In RStudio, I navigated to the correct path for this ReadMe file and opened it to edit. 
Referring to the Markdown guide, I created Headers, some italics and bold font stylings, lists, inserted a photo and a link to my ResearchGate profile. 

The assignment was relatively easy since a lot of this was simply writing out the biography in-line with the editor. The difficult area for me was correctly inserting a photo; initially did not include a full path to the file. I'm curious whether attaching a local photo has any potential problems when viewing elsewhere. When trying to get the "Preview" option to work, I encountered a few errors while trying to generate. Here, again, it fussed about my linked photo so there's some more sleuthing to do. Committing these changes to see if I'm able to see the image when viewing the ReadMe back on Github.