Updated image attempt
When viewing my ReadMe file on GitHub, I noticed that I apparently misused some of the syntax, potentially conflating some of the R Markdown conventions for regular markdown? 
I fixed these (such as using hyphens instead of asterisks for lists) and based on some suggestions online I copied the image file to the same folder as my ReadMe. Hopefully now this will work. Otherwise, the next step would be to just find an image online to link to instead of embedding with a local file.