Fixed lists
Once again, viewing on GitHub showed that the lists were incorrectly made. The syntax requires a space after dashes, OR asterisks. Referring back to the Cheatsheet online confirmed that you can use asterisks. It was just my failure to add a space that messed them up. 
I also got the preview option in R to work since the image I embedded is now in the same directory as the ReadMe file. So I was able to confirm everything looks alright this time around before pushing it off to GitHub. Third time's the charm!