Leslie Mei Homework 1
Using GitHub Desktop, I downloaded the master file for hw01. Then I right-clicked to find the file in Finder so that I could edit the correct file. I found that the most challenging part of this assignment was figuring out how the GitHub system worked - at first, I couldn't get the edits to show up because I had downloaded hw01 not through the GitHub Desktop app and edited a different version of the file in my Downloads folder. The process made a lot more sense once I thought of it in terms of syncing files with a Cloud.