LiamColes readme.md HW1
I wrote my file in R studio. Then I copied it to github, where I previewed it. The difficulties I faced were mostly in formatting the document and figuring out the process by which I upload it. The github website is not immediately intuitive, but I am slowly getting used to it. Also, I originally wanted to post a picture in the bio, but I was not able to retrieve it from my images folder when I previewed it. Hopefully, I will learn what I did wrong in the upcoming weeks.